Reorganize utility functions used in sklearn
