+++ mlflow/sklearn/__init__.py
@@ -36,6 +36,7 @@ from mlflow.utils.autologging_utils import (
     try_mlflow_log,
     INPUT_EXAMPLE_SAMPLE_ROWS,
     resolve_input_example_and_signature,
+    _get_training_session,
 )
 from mlflow.tracking._model_registry import DEFAULT_AWAIT_MAX_SLEEP_SECONDS
 
@@ -47,6 +48,7 @@ SERIALIZATION_FORMAT_CLOUDPICKLE = "cloudpickle"
 SUPPORTED_SERIALIZATION_FORMATS = [SERIALIZATION_FORMAT_PICKLE, SERIALIZATION_FORMAT_CLOUDPICKLE]
 
 _logger = logging.getLogger(__name__)
+_SklearnTrainingSession = _get_training_session()
 
 
 def get_default_conda_env(include_cloudpickle=False):
@@ -449,82 +451,6 @@ def load_model(model_uri):
     return _load_model_from_local_file(
         path=sklearn_model_artifacts_path, serialization_format=serialization_format
     )
-
-
-# NOTE: The current implementation doesn't guarantee thread-safety, but that's okay for now because:
-# 1. We don't currently have any use cases for allow_children=True.
-# 2. The list append & pop operations are thread-safe, so we will always clear the session stack
-#    once all _SklearnTrainingSessions exit.
-class _SklearnTrainingSession(object):
-    _session_stack = []
-
-    def __init__(self, clazz, allow_children=True):
-        """
-        A session manager for nested autologging runs.
-
-        :param clazz: A class object that this session originates from.
-        :param allow_children: If True, allows autologging in child sessions.
-                               If False, disallows autologging in all descendant sessions.
-
-        Example:
-
-        >>> class Parent: pass
-        >>> class Child: pass
-        >>> class Grandchild: pass
-
-        >>> with _SklearnTrainingSession(Parent, False) as p:
-        ...     with _SklearnTrainingSession(Child, True) as c:
-        ...         with _SklearnTrainingSession(Grandchild, True) as g:
-        ...             print(p.should_log())
-        ...             print(c.should_log())
-        ...             print(g.should_log())
-        True
-        False
-        False
-
-        >>> with _SklearnTrainingSession(Parent, True) as p:
-        ...     with _SklearnTrainingSession(Child, False) as c:
-        ...         with _SklearnTrainingSession(Grandchild, True) as g:
-        ...             print(p.should_log())
-        ...             print(c.should_log())
-        ...             print(g.should_log())
-        True
-        True
-        False
-
-        >>> with _SklearnTrainingSession(Child, True) as c1:
-        ...     with _SklearnTrainingSession(Child, True) as c2:
-        ...             print(c1.should_log())
-        ...             print(c2.should_log())
-        True
-        False
-        """
-        self.allow_children = allow_children
-        self.clazz = clazz
-        self._parent = None
-
-    def __enter__(self):
-        if len(_SklearnTrainingSession._session_stack) > 0:
-            self._parent = _SklearnTrainingSession._session_stack[-1]
-            self.allow_children = (
-                _SklearnTrainingSession._session_stack[-1].allow_children and self.allow_children
-            )
-        _SklearnTrainingSession._session_stack.append(self)
-        return self
-
-    def __exit__(self, tp, val, traceback):
-        _SklearnTrainingSession._session_stack.pop()
-
-    def should_log(self):
-        """
-        Returns True when at least one of the following conditions satisfies:
-
-        1. This session is the root session.
-        2. The parent session allows autologging and its class differs from this session's class.
-        """
-        return (self._parent is None) or (
-            self._parent.allow_children and self._parent.clazz != self.clazz
-        )
 
 
 @experimental
