+++ mlflow/utils/__init__.py
@@ -1,4 +1,9 @@
+from itertools import islice
+import logging
 from sys import version_info
+
+
+_logger = logging.getLogger(__name__)
 
 
 PYTHON_VERSION = "{major}.{minor}.{micro}".format(
@@ -53,3 +58,43 @@ def reraise(tp, value, tb=None):
     finally:
         value = None
         tb = None
+
+
+def _chunk_dict(d, chunk_size):
+    # Taken from: https://stackoverflow.com/a/22878842
+
+    it = iter(d)
+    for _ in range(0, len(d), chunk_size):
+        yield {k: d[k] for k in islice(it, chunk_size)}
+
+
+def _truncate_dict(d, max_key_length=None, max_value_length=None):
+    def _truncate_and_ellipsize(value, max_length):
+        return str(value)[: (max_length - 3)] + "..."
+
+    key_is_none = max_key_length is None
+    val_is_none = max_value_length is None
+
+    if key_is_none and val_is_none:
+        raise ValueError("Must specify at least either `max_key_length` or `max_value_length`")
+
+    truncated = {}
+    for k, v in d.items():
+        should_truncate_key = (not key_is_none) and (len(str(k)) > max_key_length)
+        should_truncate_val = (not val_is_none) and (len(str(v)) > max_value_length)
+
+        new_k = _truncate_and_ellipsize(k, max_key_length) if should_truncate_key else k
+        if should_truncate_key:
+            # Use the truncated key for warning logs to avoid noisy printing to stdout
+            msg = "Truncated the key `{}`".format(new_k)
+            _logger.warning(msg)
+
+        new_v = _truncate_and_ellipsize(v, max_value_length) if should_truncate_val else v
+        if should_truncate_val:
+            # Use the truncated key and value for warning logs to avoid noisy printing to stdout
+            msg = "Truncated the value of the key `{}`. Truncated value: `{}`".format(new_k, new_v)
+            _logger.warning(msg)
+
+        truncated[new_k] = new_v
+
+    return truncated
