+++ mlflow/utils/autologging_utils/__init__.py
@@ -429,3 +429,86 @@ def autologging_is_disabled(integration_name):
         return get_autologging_config(integration_name, "disable_for_unsupported_versions", False)
 
     return False
+
+
+def _get_training_session():
+    """
+    Returns a session manager for nested autologging runs.
+    """
+    # NOTE: The current implementation doesn't guarantee thread-safety, but that's okay for now
+    # because:
+    # 1. We don't currently have any use cases for allow_children=True.
+    # 2. The list append & pop operations are thread-safe, so we will always clear the session stack
+    #    once all _SklearnTrainingSessions exit.
+    class _TrainingSession(object):
+        _session_stack = []
+
+        def __init__(self, clazz, allow_children=True):
+            """
+            A session manager for nested autologging runs.
+
+            :param clazz: A class object that this session originates from.
+            :param allow_children: If True, allows autologging in child sessions.
+                                   If False, disallows autologging in all descendant sessions.
+
+            Example:
+
+            >>> class Parent: pass
+            >>> class Child: pass
+            >>> class Grandchild: pass
+
+            >>> with _SklearnTrainingSession(Parent, False) as p:
+            ...     with _SklearnTrainingSession(Child, True) as c:
+            ...         with _SklearnTrainingSession(Grandchild, True) as g:
+            ...             print(p.should_log())
+            ...             print(c.should_log())
+            ...             print(g.should_log())
+            True
+            False
+            False
+
+            >>> with _SklearnTrainingSession(Parent, True) as p:
+            ...     with _SklearnTrainingSession(Child, False) as c:
+            ...         with _SklearnTrainingSession(Grandchild, True) as g:
+            ...             print(p.should_log())
+            ...             print(c.should_log())
+            ...             print(g.should_log())
+            True
+            True
+            False
+
+            >>> with _SklearnTrainingSession(Child, True) as c1:
+            ...     with _SklearnTrainingSession(Child, True) as c2:
+            ...             print(c1.should_log())
+            ...             print(c2.should_log())
+            True
+            False
+            """
+            self.allow_children = allow_children
+            self.clazz = clazz
+            self._parent = None
+
+        def __enter__(self):
+            if len(_TrainingSession._session_stack) > 0:
+                self._parent = _TrainingSession._session_stack[-1]
+                self.allow_children = (
+                    _TrainingSession._session_stack[-1].allow_children and self.allow_children
+                )
+            _TrainingSession._session_stack.append(self)
+            return self
+
+        def __exit__(self, tp, val, traceback):
+            _TrainingSession._session_stack.pop()
+
+        def should_log(self):
+            """
+            Returns True when at least one of the following conditions satisfies:
+
+            1. This session is the root session.
+            2. The parent session allows autologging and its class differs from this session's class.
+            """
+            return (self._parent is None) or (
+                self._parent.allow_children and self._parent.clazz != self.clazz
+            )
+
+    return _TrainingSession
